+++ Bouhm.Shared/Integrations/GenericModConfigMenu/IGenericModConfigMenuApi.cs
@@ -60,6 +60,19 @@ public interface IGenericModConfigMenuApi
     /// <param name="fieldId">The unique field ID for use with <c>OnFieldChanged</c>, or <c>null</c> to auto-generate a randomized ID.</param>
     void AddNumberOption(IManifest mod, Func<int> getValue, Action<int> setValue, Func<string> name, Func<string> tooltip = null, int? min = null, int? max = null, int? interval = null, Func<int, string> formatValue = null, string fieldId = null);
 
+    /// <summary>Add a float option at the current position in the form.</summary>
+    /// <param name="mod">The mod's manifest.</param>
+    /// <param name="getValue">Get the current value from the mod config.</param>
+    /// <param name="setValue">Set a new value in the mod config.</param>
+    /// <param name="name">The label text to show in the form.</param>
+    /// <param name="tooltip">The tooltip text shown when the cursor hovers on the field, or <c>null</c> to disable the tooltip.</param>
+    /// <param name="min">The minimum allowed value, or <c>null</c> to allow any.</param>
+    /// <param name="max">The maximum allowed value, or <c>null</c> to allow any.</param>
+    /// <param name="interval">The interval of values that can be selected.</param>
+    /// <param name="formatValue">Get the display text to show for a value, or <c>null</c> to show the number as-is.</param>
+    /// <param name="fieldId">The unique field ID for use with <c>OnFieldChanged</c>, or <c>null</c> to auto-generate a randomized ID.</param>
+    void AddNumberOption(IManifest mod, Func<float> getValue, Action<float> setValue, Func<string> name, Func<string>? tooltip = null, float? min = null, float? max = null, float? interval = null, Func<float, string>? formatValue = null, string? fieldId = null);
+
     /// <summary>Add a string option at the current position in the form.</summary>
     /// <param name="mod">The mod's manifest.</param>
     /// <param name="getValue">Get the current value from the mod config.</param>